@@ -8,87 +8,84 @@ author = Holger Krekel, Bruno Oliveira, Ronny Pfannschmidt, Floris Bruynooghe, B
 license = MIT
 license_file = LICENSE
 platforms = unix, linux, osx, cygwin, win32
-classifiers = 
-	Development Status :: 6 - Mature
-	Intended Audience :: Developers
-	License :: OSI Approved :: MIT License
-	Operating System :: MacOS :: MacOS X
-	Operating System :: Microsoft :: Windows
-	Operating System :: POSIX
-	Programming Language :: Python :: 3
-	Programming Language :: Python :: 3 :: Only
-	Programming Language :: Python :: 3.5
-	Programming Language :: Python :: 3.6
-	Programming Language :: Python :: 3.7
-	Programming Language :: Python :: 3.8
-	Programming Language :: Python :: 3.9
-	Topic :: Software Development :: Libraries
-	Topic :: Software Development :: Testing
-	Topic :: Utilities
+classifiers =
+    Development Status :: 6 - Mature
+    Intended Audience :: Developers
+    License :: OSI Approved :: MIT License
+    Operating System :: MacOS :: MacOS X
+    Operating System :: Microsoft :: Windows
+    Operating System :: POSIX
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Topic :: Software Development :: Libraries
+    Topic :: Software Development :: Testing
+    Topic :: Utilities
 keywords = test, unittest
-project_urls = 
-	Source=https://github.com/pytest-dev/pytest
-	Tracker=https://github.com/pytest-dev/pytest/issues
+project_urls =
+    Changelog=https://docs.pytest.org/en/stable/changelog.html
+    Twitter=https://twitter.com/pytestdotorg
+    Source=https://github.com/pytest-dev/pytest
+    Tracker=https://github.com/pytest-dev/pytest/issues
 
 [options]
-packages = 
-	_pytest
-	_pytest._code
-	_pytest._io
-	_pytest.assertion
-	_pytest.config
-	_pytest.mark
-	pytest
-install_requires = 
-	attrs>=17.4.0
-	iniconfig
-	packaging
-	pluggy>=0.12,<1.0
-	py>=1.8.2
-	toml
-	atomicwrites>=1.0;sys_platform=="win32"
-	colorama;sys_platform=="win32"
-	importlib-metadata>=0.12;python_version<"3.8"
-	pathlib2>=2.2.0;python_version<"3.6"
-python_requires = >=3.5
-package_dir = 
-	=src
-setup_requires = 
-	setuptools>=40.0
-	setuptools-scm
+packages =
+    _pytest
+    _pytest._code
+    _pytest._io
+    _pytest.assertion
+    _pytest.config
+    _pytest.mark
+    pytest
+install_requires =
+    attrs>=19.2.0
+    iniconfig
+    packaging
+    pluggy>=0.12,<2.0
+    py>=1.8.2
+    tomli>=1.0.0
+    atomicwrites>=1.0;sys_platform=="win32"
+    colorama;sys_platform=="win32"
+    importlib-metadata>=0.12;python_version<"3.8"
+python_requires = >=3.6
+package_dir =
+    =src
+setup_requires =
+    setuptools
+    setuptools-scm>=6.0
 zip_safe = no
 
 [options.entry_points]
-console_scripts = 
-	pytest=pytest:console_main
-	py.test=pytest:console_main
+console_scripts =
+    pytest=pytest:console_main
+    py.test=pytest:console_main
 
 [options.extras_require]
-checkqa-mypy = 
-	mypy==0.780
-testing = 
-	argcomplete
-	hypothesis>=3.56
-	mock
-	nose
-	requests
-	xmlschema
+testing =
+    argcomplete
+    hypothesis>=3.56
+    mock
+    nose
+    pygments>=2.7.2
+    requests
+    xmlschema
 
 [options.package_data]
 _pytest = py.typed
 pytest = py.typed
 
 [build_sphinx]
-source-dir = doc/en/
-build-dir = doc/build
+source_dir = doc/en/
+build_dir = doc/build
 all_files = 1
 
-[upload_sphinx]
-upload-dir = doc/en/build/html
-
 [check-manifest]
-ignore = 
-	src/_pytest/_version.py
+ignore =
+    src/_pytest/_version.py
 
 [devpi:upload]
 formats = sdist.tgz,bdist_wheel
@@ -106,8 +103,3 @@ warn_return_any = True
 warn_unreachable = True
 warn_unused_configs = True
 no_implicit_reexport = True
-
-[egg_info]
-tag_build = 
-tag_date = 0
-
