@@ -50,15 +50,36 @@ props = [
   u"border-end-end-radius",
 ]
 
+els = [
+  [u'link', u'<a id="link">a</a>'],
+  [u'button', u'<button id="button">button</button>'],
+  [u'input-button', u'<input id="button-input" type="button" value="input-button">'],
+  [u'input-submit', u'<input id="submit-input" type="submit" value="input-submit">'],
+  [u'input-reset', u'<input id="reset-input" type="reset" value="input-reset">'],
+  [u'input-text', u'<input id="text-input" type="text" value="input-text">'],
+  [u'input-search-text', u'<input id="search-text-input" type="search" value="input-search-text">'],
+  [u'input-search', u'<input id="search-input" type="search" value="input-search">'],
+  [u'range', u'<input id="range-input" type="range">'],
+  [u'checkbox-input', u'<input id="checkbox-input" type="checkbox">'],
+  [u'radio-input', u'<input id="radio-input" type="radio">'],
+  [u'color-input', u'<input id="color-input" type="color">'],
+  [u'textarea', u'<textarea id="textarea">textarea</textarea>'],
+  [u'select-listbox', u'<select multiple id="select-listbox"><option>select-listbox</option></select>'],
+  [u'select-dropdown-box', u'<select id="select-dropdown-box"><option>select-dropdown-box</option></select>'],
+  [u'select-menulist-button', u'<select id="select-menulist-button"><option>select-menulist-button</option></select>'],
+  [u'meter', u'<meter id="meter" value=0.5></meter>'],
+  [u'progress', u'<progress id="progress" value=0.5></progress>'],
+]
+
 template = u"""<!-- DO NOT EDIT. This file has been generated. Source:
     ./tools/build-compute-kind-widget-fallback-props.py
 -->
 <!DOCTYPE html>
 <meta charset="utf-8">
-<title>CSS Basic User Interface Test: Compute kind of widget: {prop} disables native appearance for widgets</title>
+<title>CSS Basic User Interface Test: Compute kind of widget: {prop} disables native appearance for {el_id}</title>
 <link rel="help" href="https://drafts.csswg.org/css-ui-4/#computing-kind-widget">
 <meta name="assert" content="appropriate widget is returned when authorProps includes {prop}.">
-<link rel="match" href="../compute-kind-widget-fallback-ref.html">
+<link rel="match" href="../compute-kind-widget-fallback-{el_id}-ref.html">
 <style>
     #container {{ width: 500px; }}
     #container > #search-text-input {{ appearance: textfield; }}
@@ -66,27 +87,7 @@ template = u"""<!-- DO NOT EDIT. This file has been generated. Source:
 </style>
 
 <div id="container">
-    <a>a</a>
-    <button id="button">button</button>
-    <input id="button-input" type="button" value="input-button">
-    <input id="submit-input" type="submit" value="input-submit">
-    <input id="reset-input" type="reset" value="input-reset">
-
-    <input id="text-input" type="text" value="input-text">
-    <input id="search-text-input" type="search" value="input-search-text">
-    <input id="search-input" type="search" value="input-search">
-
-    <input id="range-input" type="range">
-    <input id="checkbox-input" type="checkbox">
-    <input id="radio-input" type="radio">
-    <input id="color-input" type="color">
-
-    <textarea id="textarea">textarea</textarea>
-    <select multiple id="select-listbox"><option>select-listbox</option></select>
-    <select id="select-dropdown-box"><option>select-dropdown-box</option></select>
-    <select id="select-menulist-button"><option>select-menulist-button</option></select>
-    <meter id="meter" value=0.5></meter>
-    <progress id="progress" value=0.5></progress>
+    {el_markup}
 </div>
 
 <script>
@@ -112,9 +113,10 @@ def write_file(path, content):
     file.write(content)
     file.close()
 
-def generate_tests(prop):
-    test = template.format(prop=prop)
-    write_file(f"kind-of-widget-fallback-{prop}-001.html", test)
+def generate_tests(prop, el_id, el_markup):
+    test = template.format(prop=prop, el_id=el_id, el_markup=el_markup)
+    write_file(f"kind-of-widget-fallback-{el_id}-{prop}-001.html", test)
 
 for prop in props:
-    generate_tests(prop)
+    for el_id, el_markup in els:
+        generate_tests(prop, el_id, el_markup)
