@@ -17,8 +17,7 @@ font_access_test(async t => {
 
   fonts.forEach(font => {
     assert_true(
-        font instanceof FontMetadata,
-        'Results should be FontMetadata instances');
+        font instanceof FontData, 'Results should be FontData instances');
 
     // Verify properties and types.
     assert_equals(typeof font.postscriptName, 'string');
@@ -31,10 +30,10 @@ font_access_test(async t => {
     assert_equals(typeof font.style, 'string', 'style attribute type');
 
     // If the sample test data contains the returned system font,
-    // then verify the values of FontMetadata.
+    // then verify the values of FontData.
     const expectedFont = testData.get(font.postscriptName);
     if (expectedFont) {
       assert_font_equals(font, expectedFont);
     }
   });
-}, 'query(): FontMetadata property types and values');
+}, 'query(): FontData property types and values');
