@@ -29,4 +29,4 @@ font_access_test(async t => {
         parsedData.tables.size, 0, "Should not have tables of size zero.");
     assert_font_has_tables(font.postscriptName, parsedData.tables, BASE_TABLES);
   }
-}, 'FontMetadata.blob(): blob has expected format and parsable table data.');
+}, 'FontData.blob(): blob has expected format and parsable table data.');
