@@ -152,11 +152,11 @@ def test_verify_payload():
     ("pr_event.json", True, {".taskcluster.yml", ".travis.yml", "tools/ci/start.sh"},
      ['lint',
       'tools/ unittests (Python 3.6)',
-      'tools/ unittests (Python 3.9)',
+      'tools/ unittests (Python 3.10)',
       'tools/ integration tests (Python 3.6)',
-      'tools/ integration tests (Python 3.9)',
+      'tools/ integration tests (Python 3.10)',
       'resources/ tests (Python 3.6)',
-      'resources/ tests (Python 3.9)',
+      'resources/ tests (Python 3.10)',
       'download-firefox-nightly',
       'infrastructure/ tests',
       'sink-task']),
@@ -189,22 +189,22 @@ def test_verify_payload():
       'wpt-firefox-stable-testharness-14',
       'wpt-firefox-stable-testharness-15',
       'wpt-firefox-stable-testharness-16',
-      'wpt-chrome-nightly-testharness-1',
-      'wpt-chrome-nightly-testharness-2',
-      'wpt-chrome-nightly-testharness-3',
-      'wpt-chrome-nightly-testharness-4',
-      'wpt-chrome-nightly-testharness-5',
-      'wpt-chrome-nightly-testharness-6',
-      'wpt-chrome-nightly-testharness-7',
-      'wpt-chrome-nightly-testharness-8',
-      'wpt-chrome-nightly-testharness-9',
-      'wpt-chrome-nightly-testharness-10',
-      'wpt-chrome-nightly-testharness-11',
-      'wpt-chrome-nightly-testharness-12',
-      'wpt-chrome-nightly-testharness-13',
-      'wpt-chrome-nightly-testharness-14',
-      'wpt-chrome-nightly-testharness-15',
-      'wpt-chrome-nightly-testharness-16',
+      'wpt-chromium-nightly-testharness-1',
+      'wpt-chromium-nightly-testharness-2',
+      'wpt-chromium-nightly-testharness-3',
+      'wpt-chromium-nightly-testharness-4',
+      'wpt-chromium-nightly-testharness-5',
+      'wpt-chromium-nightly-testharness-6',
+      'wpt-chromium-nightly-testharness-7',
+      'wpt-chromium-nightly-testharness-8',
+      'wpt-chromium-nightly-testharness-9',
+      'wpt-chromium-nightly-testharness-10',
+      'wpt-chromium-nightly-testharness-11',
+      'wpt-chromium-nightly-testharness-12',
+      'wpt-chromium-nightly-testharness-13',
+      'wpt-chromium-nightly-testharness-14',
+      'wpt-chromium-nightly-testharness-15',
+      'wpt-chromium-nightly-testharness-16',
       'wpt-chrome-stable-testharness-1',
       'wpt-chrome-stable-testharness-2',
       'wpt-chrome-stable-testharness-3',
@@ -258,11 +258,11 @@ def test_verify_payload():
       'wpt-firefox-stable-reftest-3',
       'wpt-firefox-stable-reftest-4',
       'wpt-firefox-stable-reftest-5',
-      'wpt-chrome-nightly-reftest-1',
-      'wpt-chrome-nightly-reftest-2',
-      'wpt-chrome-nightly-reftest-3',
-      'wpt-chrome-nightly-reftest-4',
-      'wpt-chrome-nightly-reftest-5',
+      'wpt-chromium-nightly-reftest-1',
+      'wpt-chromium-nightly-reftest-2',
+      'wpt-chromium-nightly-reftest-3',
+      'wpt-chromium-nightly-reftest-4',
+      'wpt-chromium-nightly-reftest-5',
       'wpt-chrome-stable-reftest-1',
       'wpt-chrome-stable-reftest-2',
       'wpt-chrome-stable-reftest-3',
@@ -280,8 +280,8 @@ def test_verify_payload():
       'wpt-servo-nightly-reftest-5',
       'wpt-firefox-stable-wdspec-1',
       'wpt-firefox-stable-wdspec-2',
-      'wpt-chrome-nightly-wdspec-1',
-      'wpt-chrome-nightly-wdspec-2',
+      'wpt-chromium-nightly-wdspec-1',
+      'wpt-chromium-nightly-wdspec-2',
       'wpt-chrome-stable-wdspec-1',
       'wpt-chrome-stable-wdspec-2',
       'wpt-webkitgtk_minibrowser-nightly-wdspec-1',
@@ -289,12 +289,12 @@ def test_verify_payload():
       'wpt-servo-nightly-wdspec-1',
       'wpt-servo-nightly-wdspec-2',
       'wpt-firefox-stable-crashtest-1',
-      'wpt-chrome-nightly-crashtest-1',
+      'wpt-chromium-nightly-crashtest-1',
       'wpt-chrome-stable-crashtest-1',
       'wpt-webkitgtk_minibrowser-nightly-crashtest-1',
       'wpt-servo-nightly-crashtest-1',
       'wpt-firefox-stable-print-reftest-1',
-      'wpt-chrome-nightly-print-reftest-1',
+      'wpt-chromium-nightly-print-reftest-1',
       'wpt-chrome-stable-print-reftest-1'])
 ])
 def test_schedule_tasks(event_path, is_pr, files_changed, expected):
