@@ -60,6 +60,6 @@ def load_product_update(config, product):
     data = module.__wptrunner__
 
     update_properties = (getattr(module, data["update_properties"])()
-                         if "update_properties" in data else {})
+                         if "update_properties" in data else (["product"], {}))
 
     return update_properties
