@@ -7,7 +7,7 @@ from . import assert_javascript_entry
 
 
 @pytest.mark.asyncio
-async def test_types_and_values(bidi_session, current_session, inline, wait_for_event):
+async def test_types_and_values(bidi_session, current_session, current_time, inline, wait_for_event):
     await bidi_session.session.subscribe(events=["log.entryAdded"])
 
     on_entry_added = wait_for_event("log.entryAdded")
@@ -15,7 +15,7 @@ async def test_types_and_values(bidi_session, current_session, inline, wait_for_
     expected_text = current_session.execute_script(
         "const err = new Error('foo'); return err.toString()")
 
-    time_start = math.floor(time.time() * 1000)
+    time_start = current_time()
 
     # TODO: To be replaced with the BiDi implementation for navigate.
     current_session.url = inline(
@@ -23,7 +23,7 @@ async def test_types_and_values(bidi_session, current_session, inline, wait_for_
 
     event_data = await on_entry_added
 
-    time_end = math.ceil(time.time() * 1000)
+    time_end = current_time()
 
     assert_javascript_entry(
         event_data,
